+++ src/ResourceManager/Sql/Commands.Sql/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Current Release
 * Updated documentation of -ComputeGeneration parameter in several cmdlets
+* Adding new Cmdlets for Management.Sql to allow customers to add TDE Certificate to Sql Server instance or a Managed Instance
+	- Add-AzureRmSqlServerTransparentDataEncryptionCertificate
+	- Add-AzureRmSqlManagedInstanceTransparentDataEncryptionCertificate
 
 ## Version 4.6.1
 * Updated example in the help file for Get-AzureRmSqlDatabaseExpanded